changed s_a and sum_s_a data types to long double (float and double were not large enough)

+++ src/calc_n_a/clone.h
@@ -15,7 +15,7 @@ typedef struct Clone_s
 
 	float wDotIntegral = 0.0 ;
 
-	double s_a ;
+	long double s_a ;
 
 	int n_a ;
